game_project/backend/requirements.txt

Django==3.2   
redis==4.0.2   
djangorestframework==3.12.4 
channels
channels-redis
python-dotenv
dj-database-url
psycopg2
djangorestframework
daphne
redis 
django-redis
celery[redis]
djangorestframework-simplejwt
django-cors-headers
geoip2



